requirements.txt
streamlit==1.37.1
streamlit-option-menu==0.3.13
tensorflow==2.17.0
pillow==10.4.0
numpy==1.26.4
matplotlib==3.9.2
